move to different file, resolve circular import
