+++ src/allotropy/allotrope/exceptions.py
@@ -1,0 +1,3 @@
+
+class AllotropeConversionError(Exception):
+    pass
