+++ src/allotropy/allotrope/models/shared/definitions/definitions.py
@@ -4,7 +4,7 @@ from dataclasses import dataclass
 from enum import Enum
 from typing import Optional, Union
 
-from allotropy.allotrope.allotrope import AllotropeConversionError
+from allotropy.allotrope.exceptions import AllotropeConversionError
 
 TTupleData = list[Optional[Union[float, bool, str]]]
