+++ src/allotropy/allotrope/schema_parser/model_class_editor.py
@@ -7,7 +7,7 @@ from pathlib import Path
 import re
 from typing import Optional
 
-from allotropy.allotrope.allotrope import AllotropeConversionError
+from allotropy.allotrope.exceptions import AllotropeConversionError
 from allotropy.allotrope.schema_parser.schema_model import (
     get_all_schema_components,
     get_schema_definitions_mapping,
