+++ src/allotropy/parsers/agilent_gen5/plate_data.py
@@ -11,7 +11,7 @@ from typing import Optional, Union
 import numpy as np
 import pandas as pd
 
-from allotropy.allotrope.allotrope import AllotropeConversionError
+from allotropy.allotrope.exceptions import AllotropeConversionError
 from allotropy.allotrope.models.shared.definitions.definitions import (
     FieldComponentDatatype,
     TDatacube,
