+++ src/allotropy/parsers/appbio_quantstudio/appbio_quantstudio_builders.py
@@ -8,7 +8,7 @@ import uuid
 import numpy as np
 import pandas as pd
 
-from allotropy.allotrope.allotrope import AllotropeConversionError
+from allotropy.allotrope.exceptions import AllotropeConversionError
 from allotropy.allotrope.models.pcr_benchling_2023_09_qpcr import (
     ExperimentType,
 )
