+++ src/allotropy/parsers/appbio_quantstudio/views.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 from collections.abc import Iterator
 from typing import Generic, Optional, TypeVar, Union
 
-from allotropy.allotrope.allotrope import AllotropeConversionError
+from allotropy.allotrope.exceptions import AllotropeConversionError
 
 T = TypeVar("T")
