+++ src/allotropy/parsers/example_weyland_yutani/example_weyland_yutani_parser.py
@@ -1,7 +1,7 @@
 from io import IOBase
 import uuid
 
-from allotropy.allotrope.allotrope import AllotropeConversionError
+from allotropy.allotrope.exceptions import AllotropeConversionError
 from allotropy.allotrope.models.fluorescence_benchling_2023_09_fluorescence import (
     ContainerType,
     DeviceControlAggregateDocument,
