+++ tests/parsers/roche_cedex_bioht/roche_cedex_bioht_structure_test.py
@@ -1,7 +1,7 @@
 import pandas as pd
 import pytest
 
-from allotropy.allotrope.allotrope import AllotropeConversionError
+from allotropy.allotrope.exceptions import AllotropeConversionError
 from allotropy.allotrope.models.shared.definitions.custom import (
     TNullableQuantityValueMilliOsmolesPerKilogram,
 )
